% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_1x1.R
\name{data_1x1}
\alias{data_1x1}
\title{Repeated measures data structure}
\usage{
data_1x1(array_1, array_2, jit_distance = 0.09, jit_seed = 321)
}
\arguments{
\item{array_1}{<numeric> the array of datapoints to be plotted}

\item{array_2}{<numeric> the array of datapoints to be plotted}

\item{jit_distance}{<numeric> the amount of distance between jittered datapoints (0 by default)}

\item{jit_seed}{<numeric> the amount used in set.seed() (321 by default)}
}
\value{
data_1x1
}
\description{
This function transforms your data array('s) into the data structure that is needed to visualize a 1x1 repeated measure.
}
\examples{
\dontrun{
# Using an example dataset
data_1x1 <- data_1x1(
                    array_1 = Sepal.Length[1:50],
                    array_2 = Sepal.Length[51:100],
                    jit_distance = .09,
                    jit_seed = 321)

}
}
