% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raincloud_1x1.R
\name{raincloud_1x1}
\alias{raincloud_1x1}
\title{1 x 1 comparison}
\arguments{
\item{data_1x1}{<data.frame> the array of datapoints to be plotted}

\item{colors}{<string> concatenated string for both colors}

\item{fills}{<string> concatenated string both both fills}

\item{size}{<numeric> data size}

\item{alpha}{<numeric> data alpha}

\item{ort}{<string> vertical or horizontal display of plot}
}
\description{
This function visualizes a 1x1 repeated measures raincloud.
}
\examples{
\dontrun{
# Using an example dataset
raincloud_1x1 <- function(data_1x1,
                         colors = (c('dodgerblue','darkorange')),
                         fills = (c('dodgerblue', 'darkorange')),
                         size = 1.5,
                         alpha = .6,
                         ort = 'v')

}
}
