% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raincloud_1x1_repmes.R
\name{raincloud_1x1_repmes}
\alias{raincloud_1x1_repmes}
\title{1 x 1 repeated measures raincloud}
\arguments{
\item{data_1x1}{<data.frame> the array of datapoints to be plotted}

\item{colors}{<string> concatenated string for both colors}

\item{fills}{<string> concatenated string for both fills}

\item{line_color}{<string> color of the lines}

\item{line_alpha}{<numeric> alpha of the lines}

\item{size}{<numeric> data size}

\item{alpha}{<numeric> data alpha}

\item{align_clouds}{<bool> FALSE if spreaded, TRUE if aligned}
}
\description{
This function visualizes a 1x1 repeated measures raincloud.
}
\examples{
\dontrun{
# Using an example dataset
raincloud_1x1_repmes <- function(data_1x1,
                                 colors = (c('dodgerblue','darkorange')),
                                 fills = (c('dodgerblue', 'darkorange')),
                                 line_color = 'gray',
                                 line_alpha = .3,
                                 size = 1.5,
                                 alpha = .6,
                                 align_clouds = FALSE)

}
}
