% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raincloud_2x2_repmes.R
\name{raincloud_2x2_repmes}
\alias{raincloud_2x2_repmes}
\title{2 x 2 repeated measures raincloud}
\arguments{
\item{data_2x2}{<data.frame> the array of datapoints to be plotted}

\item{colors}{<string> concatenated string for both colors}

\item{fills}{<string> concatenated string for both fills}

\item{line_color}{<string> color lines}

\item{line_alpha}{<numeric> alpha lines}

\item{size}{<numeric> data size}

\item{alpha}{<numeric> data alpha}

\item{spread_x_ticks}{<bool> TRUE if 4 x ticks, FALSE if 2 x ticks}
}
\description{
This function visualizes a 2x2 repeated measures raincloud.
}
\examples{
\dontrun{
# Using an example dataset
raincloud_2x2_repmes <- function(data_2x2,
                                colors = (c('dodgerblue', 'darkorange',
                                'dodgerblue', 'darkorange')),
                                fills = (c('dodgerblue', 'darkorange',
                                'dodgerblue', 'darkorange')),
                                line_color = 'gray',
                                line_alpha = .3,
                                size = 1.5,
                                alpha = .6,
                                spread_x_ticks = TRUE)

}
}
