% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raincloud_2x3_repmes.R
\name{raincloud_2x3_repmes}
\alias{raincloud_2x3_repmes}
\title{2 x 3 repeated measures raincloud}
\arguments{
\item{data_2x2}{<data.frame> the array of datapoints to be plotted}

\item{colors}{<string> concatenated string of the colors}

\item{fills}{<string> concatenated string of the fills}

\item{size}{<numeric> data size}

\item{alpha}{<numeric> data alpha}
}
\description{
This function visualizes a 2x3 repeated measures raincloud.
}
\examples{
\dontrun{
# Using an example dataset
raincloud_2x3_repmes <- function(data_2x2,
                                 colors = (c('dodgerblue, 'darkorange', 'dodgerblue',
                                 'darkorange', 'dodgerblue', 'darkorange')),
                                 fills = (c('dodgerblue', 'darkorange', dodgerblue',
                                 'darkorange', 'dodgerblue', 'darkorange')),
                                 size = 1.5,
                                 alpha = .6,
                                 ort = 'v')
}
}
